profile::docker::storage::loopback: remove from puppet
The result of this removal is that we will switch to running docker using
the overlay2 driver. Overlay2 is the recommended storage driver for
docker since well before we switched to debian buster for the build
servers, the only ones using this class. We'll "lose"
all the docker data on the build hosts, both in production and on
wmcs, but the impact should be limited to those.

Change-Id: Ie3754806cb15dfc8f19a8ca194a0f742ba63e579