sonofgridengine: grid_configurator: wrap subproccess calls
Bofore this patch, if any of the internal grid command failed, the admin would
not see the stderr.

This patch introduces a function to wrap all subproccess calls with improved
error reporting.

While at it, some calls are refreshed, removing things like check=True which could
halt the execution of the configurator, leaving config in bad state.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I9971ca91810dec79e7b5d4a5c835e1ac116e3fcf