add parsoid-rt-tests.wikimedia.org to alternate_domains
As suggested by Majavah adding the domain to this list
hoping it means the "if (req.url ~ "^/static/")" is disabled in varnish.

To fix T266509 which caused us some headaches why there were "mystery
404s" but only for files under a ./static/ URL while everything else
worked.

Also, I set it to "pass", so no caching. The other option would be
"normal" but not 100% sure which one subbu and team prefer.

Bug: T266509
Change-Id: Ic59fc8ff1fb7711bbcc0244ebe87be3b82f376a0