phabricator: move vcs firewall rules to profile
This is just a style change to satisfy the "no ferm rules
in the middle of modules, only in profiles" rule.

Separately it is questionable if we really need to have this
array anymore and git sshd should listen on anything besides
the actual git-ssh IP. But also we still want to support
both IPv4 and IPv6.

The intention right now is not to change any of that given
T296022 already discusses deprecation of the entire service.

Bug: T114209
Bug: T296022
Change-Id: I8e02117bb36a17ef3883e216d2f8bb5a432792a5