envoy: make the choice of api version explicit
We need to make the choice consistently whenever we have multiple
listeners (for instance, when we have both TLS termmination and services
proxies in the same place).

So, introduce an api_version parameter in profile::envoy and propagate it
everywhere that class is used.

A subsequent patch will introduce v3 support for the services proxy as well.

Change-Id: I87897c49bc423ac9b0abf94c0486233de9d79a5e