icinga: let Jack Gleeson run commands for any host or service
Jack is in fundraising-tech-ops and has existing shell access with
analytics-privatedata-user.

He can't run commands (like ACK, downtime, etc) on Icinga alerts.

The proper way to do this would be to add a contact to the fr-tech-ops
Icinga contact group and use that contact group for all the fundraising
related monitoring hosts and services.

First step was a missing personal icinga contact in the private repo.
I added that.

But then it wasn't just adding to the existing contact group because
that is not actually being used (!?).

A "grep contact_groups puppet_services.cfg | sort | uniq" shows not
a single case of the fr-related contact groups, even though the groups
themselves exist.

So to fix this regardless without touching all the existing checks
adding Jack to global privileges in Icinga's cgi.cfg where global
overrides for any host or service exist.

Adding alongside other SREs there.

Bug: T298649
Change-Id: I4ecebe176cd9f82748c0dceeb2a933ff34ed8770