Use a flag to deploy log4j-extras on Hadoop-related nodes
In T276906 we added the log4j-extras jar to the Hadoop classpath
to enable the rolling gzip file appender, that helped a lot
in reducing the size of the logs (masters and workers).
After the recent changes to Hive's log4j
(see https://issues.apache.org/jira/browse/BIGTOP-3619),
it became clear that the extra jar clashes with more recent
log4j versions (at least in the context of Hive's Datanucleus
4.1.x code). This change wraps the Hadoop setting with a hiera
flag, in order to selectively deploy it. By default we'll keep
it enabled, and for the moment we'll disable it only where
hive daemons (like hive server/metastore) run in Hadoop test.

Change-Id: I869730cee6f1784b6f246023cc397b654bde8a03