role::kafka::main: apply kafka fixed uid/gid to eqiad
This is a follow up to move the Kafka main eqiad cluster
to the fixed kafka uid/gid scheme. The change has already been applied
in codfw and it worked nicely during reimages.
This change needs to be merged with puppet disabled, and every host
needs to be moved to the new uid/gid manually (via script) beforehand.

Bug: T296641
Change-Id: I215044dcde661734f51923850bfa349bbc9a786d