Add Elliot Eggleston (ejegg) to fr-tech-ops Icinga contact group.
This is on the back of a task Jack Gleeson raised when he was unable
to ACK alerts in Icinga.  That was resolved but the same lack of
permissions applies to all of fundraising so he has provided a list
of users who need the same change made.

The fix for Jack added him both to the fr-tech-ops contact group, and
added his user directly to the "authorized_for_system_information",
"authorized_for_all_service_commands" and
"authorized_for_configuration_information" in-built Icinga groups.

Daniel who completed this change had initially only done the latter,
so there is an open question on whether simply adding to the
fr-tech-ops contact group would have been sufficient.

To assess this I am adding one of the users, ejegg, to the fr-tech-ops
contact group and we will see if they have the required permission
without adding to the list of users who are globally authorized.

Bug: T298649
Change-Id: I9c30b7b7d5c4b028dc4482e29fa5f1873ba9bc83