kafka: add check to test the Broker's TLS port
We are currently missing checks to ensure that the Kafka brokers
are running a valid TLS certificate. All the clusters are using
the Puppet host TLS certificates, and they have a loong expiry,
but we are in the process of moving all the clusters to the Kafka
PKI intermediate and a good safety net is needed.
For example, me and Ben investigated a problem on the test cluster
that was very sneaky. The test cluster runs with PKI certificates,
that by default have a very short expiry time (some months, but
hopefully tunable). The certificates expired in December, newer
one were provisioned on the hosts by puppet but since we didn't
restart kafka, the old ones kept being used. The cluster ran fine
until the kafka broker acting as controller got restarted, triggering
new TLS connections between brokers that suddently started to
fail one after the other.

To better support this change:
- added a new value to kafka_config to expose the ssl port used.
- added a special check_ssl_kafka nagios command that uses
  the --no-verify option (needed until we migrate all clusters
  away from the current cergen certs).

Change-Id: I5a51f8c18b172a54fdde3f7edc6aa19ccc011435