P:admin: switch to using wmflib::dump_params
This CR should also be used to evaluate the implementation of dump_params.
dump_params is already merged but only used in cloud currently

This CR also adds some tests further it changes the implmentation so
that it works with ruby2.4.  The puppet masters in production or 2.5 so
this shouldn't be strictly needed but for now it keep it inline wit a PR
in extlib[1]

 [1]https://github.com/voxpupuli/puppet-extlib/pull/194

Hosts: sretest1001.eqiad.wmnet
Hosts: sretest1002.eqiad.wmnet
Hosts: puppetmaster1001.eqiad.wmnet
Change-Id: Ic7fcac8206d9155cdff0011ffb19d52dda9d7b6a