partman: use reuse profiles on all restbase hosts
Allow us to reimage restbase hosts without losing instance-data or the
JBOD volumes on all restbase hosts. This will only impact hosts when
we explicitly reimage them. Once this is merged I will test reimaging
restbase2010 without decommissioning to ensure that it comes back up
and rejoins the cluster safely, and then move on to reimaging all on
buster.

Bug: T295375
Change-Id: I171b4ac60cff877b884e3c1eefa2d681b1218e75