Atlas exporter: add probes and traceroute mesurements
Add ripe probes to anchors mesurements to our existing achors to anchors.
This for example adds 1700 v4 pings and 1200 v6 pings to eqiad anchors.
More than doubling the existing.
Probes are ran from end user networks, so they can be more flaky, but
also more representative of end user experience.

Add traceroute mesurements, it will be useful to graph the number of
hops to see upstreams events.

eqiad and codfw only for now, will add the other sites if successful.

Not sure why "-streaming" is curently set to false and if we will hit
some limits (on the Atlas or on the Prometheus side).

Bug: T251156
Change-Id: I9f75a02b53f12353d6a8c4dad54ec101bbc609d1