role::pki::multirootca: add dedicated profile for ml-serve k8s
The ml-serve eqiad/codfw clusters are going to use cert-manager
and cfssl-issuer to retrieve .discovery.wmnet certificates
(and possibly more in the future, but we have a limited scope
for the moment). This change adds a dedicated profile as it was
done for k8s_staging (no need for a restricted/different expiry
setup).

Bug: T298976
Change-Id: I282c5b48447ea41bae2ae233cc60daaaf5729fc2