P:microsites::os_reports: update rsync to use -r instead of *
The asterisk can cause confusion in rsync commands as its not always
clear if it is interpreted as a glob by the shell or passed to the rsync
command.  In this case the asterisk works as intended however lets be
explicit and use -r to avoid confusion.  I have also added -t to
preserve the original times

Hosts: P:microsites::os_reports
Bug: T299520
Change-Id: Ib3480423e4e214565de0dede82fbc21ae3d55326