Stop writing parquet logs to files
We have seen an issue with the use of the FileHandler logging for
parquet, which sometimes causes users' queries to fail owing to buggy
behaviour of the %u parameter.

For this reason we are reverting to using the ConsoleHandler for logging
and reducing the verbosity of logs, so that only messages that are
SEVERE will be logged.

Bug: T297734
Change-Id: I797b9b2378db23d80d05928b04bdc131c7986c86