ci: set Docker partition size explicitly
/var/lib/docker is created first consuming 70% of the disk space
available which turns to 42G for a disk80 instance. That is a lot of
space since we typically use 15G-20G here.

Set the partition size explicitly to 24G which is enough to fit our
Docker images.

To do the migration for an agent:

- https://integration.wikimedia.org/ci/computer/
- Put agent offline
- Wait for jobs to complete
- Disconnect agent

On the agent:

  systemctl stop docker
  umount /var/lib/docker
  umount /srv

If files are busy check `lsof /var/lib/docker /srv`, kill process and
try again.

Then remove the logical volumes:

  lvremove -y /dev/vd/docker
  lvremove -y /dev/vd/second-local-disk

Run Puppet.

It would recreate the partition, populate some base directories and
clone the mirrored git repositories.  Check partitions are ok with `df`
they should be:

24G /var/lib/docker
37G /srv

Start docker and check it:

  systemctl start docker
  docker pull docker-registry.wikimedia.org/releng/ci-buster

Reconnect the agent in Jenkins.

Bug: T292729
Change-Id: I8ffe2cc80fb9d4fc23ed4cda0a436a1ee67b9755