Use the default prometheus_mysql_exporter for matomo
The mariadb service on this has was previously converted from a
multi-instance to a single-instance configuration. However, the
prometheus_mysql_exporter continued to use a named configuration,
different from the default.

This continued to work for a while, but recently stopped working to to
improvements made to the monitoring configuration for MariaDB. Matrics
from matomo ceased to be collected as a result.

This commit updates the matomo configuration so that it correctly
configures the default prometheus_mysql_exporter and updates the scrape
target accordingly.

Bug: T299762
Change-Id: If1cabdf32327cae4af6f46401c1568ca4f3497e8