openstack: tell dynamicproxy to not edit dns records
I want to move DNS handling to the dynamicproxy API so that we can give
API access to edit dynamicproxy proxies without giving that client
access to manage the entire wmcloud.org. DNS zone.

To do that, we set a specific HTTP header on all API requests that
already deal with DNS records so that we can later tell the dynamicproxy
api to avoid duplicate updates when that header is set.

Bug: T295246
Change-Id: Ia4295b9973d5098eef039eecbfd007bfa34b150c