Add a kafka_11 profile to the PKI Kafka Intermediate settings
It seems that our version of Kafka, 1.1, doesn't allow the dynamic
reload of keystores with the same filepath/password
(see https://issues.apache.org/jira/browse/KAFKA-7429).
To avoid a roll restart of the clusters using PKI every month
(for the moment, only kafka test) expand the default cert validity
to a year.
The idea is to remove this constraint when we'll migrate to a new
version of Kafka.

Bug: T299409
Change-Id: Iff7d7168fb8afb484bb610efe586ab0ae3e7038e