install_server: swift UID/GID should match filesystems (if present)
To make reimages easier (e.g. for distro upgrades), we want to ensure
that, on backend nodes with existing swift filesystems, that the swift
user uid and gid match those already on the filesystem. It would be
very time-consuming to chown all the files.

So the swift user/group creation is parameterised, defaulting to 902
(the fleet-wide setting we are moving towards).

If /dev/sde1 is mountable as an xfs filesystem (it is present thus on
all the backend nodes), then it is mounted to a temporary directory
(to avoid mountpoint-clash) and `stat` is called on "objects" (a
directory present at the top level of sde1 on all the backend ndoes)
to find the uid and gid to use for the swift user.

late_command.sh is run in the installer (which runs dash & busybox),
which gives us a rather restricted set of commands to work with -
hence not trying to use lsblk or label-based mounting. This also means
we can't use `trap ERR` to unmount the filesystem on error, so we have
to ensure that none of the pipelines while it is mounted can cause the
script to terminate.

Bug: T300057
Change-Id: Iff4771654f7597db239bfa8a43da9cb84067cf97