rdf query service: Use constant filename for defaults
Parameterizing this filename means that the script using it
needs to know the parameterization. We don't intend to deploy
wcqs and wdqs updaters to the same host, so make things easy
by using a generic name for the file.

Dependent patch must be merged and a deploy of the query service
to wdqs and wcqs must be run prior to merging this patch. After
deployment the old defaults files, /etc/default/w{c,d}qs-updater,
should be deleted from all instances to avoid confusing future
debugging efforts.

Bug: T299222
Depends-On: Iddc00057f1a33df852bf807b4a522ac9c28a1e89
Change-Id: I5c413529e6a6d10cc420fe947b7c42f3dbb3b10f