maps: tweak postgres configuration settings to use more resources
We're hitting issues with connection contention when we do imports,
and we are probably not getting the best performance out of the
connections we do make. currently postgresql instances do not use a
lot of memory or CPU, and they should be allowed to use more
particularly in light of better specs for the maps hosts.

Increase wal buffers because we have the resources and this will help
our replicas

Test with a single instance first

Bug: T298246
Change-Id: Iada64fe1c94c9452a9effa471ba12b14bfcc357a