zuul: prune heads and tags on each fetches
This should improve the zuul-merger operations. When it resets a
repository, it recreates local branches from origin/ which keep
accumulating since we never prune them.  One of the merge has 130
branches for mediawiki/core!

Similarly, keep refs/tags/ in sync as well for good measure. That might
have made T252310 a non issue when it tries to fetch a tag update which
is forbidden by default since git 2.20.

Bug: T220606
Bug: T252310
Change-Id: I2f2046f779535f7310542a49d31057fc1ef4a344