Add hostname-override and cluster-cidr to kube-proxy
Kube-proxy needs to know the hostname kubelet used to register the local
node as. If it's overridden in kubelet, it must be overridden in
kube-proxy as well.

Also provide the cluster-cidr to kube-proxy so it can distinguish
between internal and external traffic. This does only support IPv4 for
now as we're not running full dual stack currently (and k8s 1.16 does
not fully support it).

Bug: T300500
Change-Id: Ie3db31d030d7d7b674a097a44767311a0d4e660e