ci: Qemu image and snapshot creation
This is intended to automatize a manual runbook we use to create a
virtual machine base image. The use case is to run arbitrary Docker
image to test fresh and mw-cli.

== context ==
The image creation process got started by Timo and is described at:
https://www.mediawiki.org/wiki/Continuous_integration/Qemu

The instance holding the images is integration-agent-qemu-1001 which is
running Stretch. We have created a Bullseye replacement
integration-agent-qemu-1003.

Dave Pifke replayed the runbook in November 2021 (T284774) but
unfortunately I have repartitionned the instance late in January 2022
and wiped all the /srv partition which was holding the images.

== details ==

The introduced script downloads Debian official cloud image, creates a
copy of the qcow2 image with extended disk space then adds a qcow2 layer
that is customized to include a ssh key and the few packages we need.

I went to use libguestfs-tools set of utilities since they are well
documented and straightforward to use. The image customization is done
with just a few options passed to `virt-customize`.

The script invoked by Puppet is reentrant and once it has completed once
would be very quick to execute on each puppet runs (just a few file
existence tests are made).

Bug: T284774
Change-Id: I3a43d2ae5a5c119daf6b9b922bc4ae0cb0b7e2e4