gitlab: parameter for exporters expects Hash but is array by default
Noticed in cloud:

Error while evaluating a Function Call, Class[Profile::Gitlab]:
parameter 'exporters' expects a Hash value, got Array

because the parameter here has a data type of Hash but is an array by
default.

[] -> {}

Doesn't happen in prod because we set a specific value there
and don't fall back to default.

Bug: T297411
Change-Id: I95bf372c3e9f54383c26e0b0dcaf4a09e2becf82