Use the system default mysql prometheus exporter for analytics-meta and matomo
Both the matomo and analytics-meta databases use the system default
systemd unit file to manage their mariadb instance.

This means that they now need to use the system default prometheus
exporter for mariadb as well, rather than a named instance.

The port used by the exporter also reverts to the default of 9104.

Bug: T299762
Change-Id: I3680e6f3789d9b6db39d65f34672d4c1e94cbb2e