gitlab: parameter to allow using either acmechief or certbot for certs
This is to work around the fact that acmechief won't work in cloud
unless you setup your own acmechief server for just your
cloud VPS project.

Copying the setup from Gerrit in devtools.

If in cloud we will just install certbot and use that.

Either way we are ending up with Letsecnrypt certs.

Explicitly turning it on/off in both environments with comments.

Reusing the "/usr/bin/gitlab-ctl hup nginx" command to refresh
nginx after cert renewal with certbot.

Bug: T297411
Change-Id: I724e7dc7e7dab7b2a578fa0d35a58ad142fa7667