planet: add diff.wikimedia.org/feed to en.planet.wikimedia.org
https://en.planet.wikimedia.org is the English language
version of our Wikimedia feed aggregator service that has been
around for many years and turns a lot
of Wikimedia related blogs and feeds into a single one.

So I thought this would be interesting for movement comms.

After hearing the section about http://diff.wikimedia.org
on today's WMF staff meeting I wanted to check if
diff provides a feed and we already include that in the
planet feed.

Turns out, yes, it has a feed, because it is Wordpress
and every Wordpress just comes with the /feed/ URL.

So this includes it here, but just for the English
version of planet. Note we have other language versions
as well. So the best outcome would be that we link
the different language versions of diff you mentioned
to the according language version of planet.

Either way this is like cross-linking and should increase
your reach in a very Wikimedia focused audience that has
been around for a long time.

https://wikitech.wikimedia.org/wiki/Planet.wikimedia.org

https://meta.wikimedia.org/wiki/Planet_Wikimedia

Bug: T230444
Change-Id: I582d5bf3cb3fd5860b983805c85c5f196d68b732