profile::docker::engine: add param to ignore docker storage settings
To ease the migration to overlayfs and avoid host conditionals
in roles, it seems more flexible to add a specific hiera option
to ignore docker storage class settings (mostly devicemapper-related).

The idea is to turn the flag "true" before reimaging a node,
so that its first puppet run will be deploy Docker overlayfs
settings.

Bug: T300744
Change-Id: Iea3a37c22cceee4849c003d6a718447689ce4b46