admin: make Tyler Cipriani the approver for the deployment group
Tyler has been approving access requests to be added to the
deployment group on a regular basis in the past.

Also currently another one is in progress at T300664.

But in code we did not have an approver line for this group yet
which lead to questions on the ticket who should approve or if one
is needed.

At the same time we do have the approver line for the "restricted"
group right below and the approver is Tyler.

The "restricted" group is a subset of deployment.

So here I suggest to also make that official for the deployment group.

I am thinking that maybe that even is just a "bug"/unintentional that
it was only applied to "restricted" but not here so far.

I wanted to keep this question seperate from the ongoing access request
for a specific person though.

Bug: T300664
Change-Id: I7cc5ba06c7de017f092d6598ddac1ba09ce56f91