kubeadm: node-upgrade: ignore empty lines in host list
Updates the wmcs-k8s-node-upgrade script to ignore empty lines in the
given host file. This is primarly intended to make using the script less
annoying with a text editor that inserts an empty line at the end of the
file. While at it, changed line[:-1] (to remove the trailing end of line)
to line.strip() to remove all kinds of whitespace.

Change-Id: Ia5c91d45a65ae3ee961485020b5b15b3d0d47f41