toolforge: grid: always install latest jobutils package
I found a hard-to-debug problem. A tool user submitting scheduled jobs using
cron, the cron server had an older version of jobutils, which didn't
support the '-release' option, resulted in the job being scheduled in the wrong
grid.

Real life example:

  ===== NODE GROUP =====
  (2) toolsbeta-sgegrid-master.toolsbeta.eqiad1.wikimedia.cloud,[...]
  ----- OUTPUT of 'dpkg -s jobutils... version || true' -----
  Version: 1.41
  ===== NODE GROUP =====
  (1) toolsbeta-services-01.toolsbeta.eqiad1.wikimedia.cloud
  ----- OUTPUT of 'dpkg -s jobutils... version || true' -----
  Version: 1.33
  ===== NODE GROUP =====
  (10) toolsbeta-sgebastion-[04-05].toolsbeta.eqiad1.wikimedia.cloud,[...]
  ----- OUTPUT of 'dpkg -s jobutils... version || true' -----
  Version: 1.43
  ===== NODE GROUP =====
  (3) toolsbeta-sgeexec-[0901-0902].toolsbeta.eqiad1.wikimedia.cloud,[...]
  ----- OUTPUT of 'dpkg -s jobutils... version || true' -----
  Version: 1.42

We usually have controlled (and well tested) releases of this package.
For misctools we already use 'latest' as well.

So, change to latest and see if we trully have good practices when releasing
the package.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I75f4b79b14eb5923b0ef312e24220c4081676abf