etherpad: allow setting listening IP in Hiera, use IPv6 on etherpad1003
When trying to move Etherpad from etherpad1002 to etherpad1003 we ran
into a problem:

envoy tries to connect to etherpad using IPv6, which fails because
etherpad isn't listening on it and never has, but also envoy does not
fall back transparently to IPv4 for some reason, which was unexpected.

On etherpad1002 this was not an issue because it does not have a AAAA
record unlike etherpad1003.

The first impulse was to change the envoy config, which uses
etherpad1003.eqiad.wmnet, and use something like 127.0.0.1 instead.

But on second thought, why not make Etherpad actually listen on IPv6.

It turns out that if you put either an empty value or "::" as value
for "ip" into the Etherpad config it will listen on IPv6. (tested in
cloud).

It is ONLY IPv6 though, not both, despite of some online sources
claiming it should listen on BOTH if you set an empty value.
This depends on the node version though as well.

Anyways, if it is only IPv6 and works.. that should be just fine since
this is behind caching layer anyways.

Since the etherpad module also already has this as a parameter just
need to pass it through from the profile and add a Hiera lookup.

Then we can leave it unchanged on etherpad1002 and try it on
etherpad1003.

Bug: T300568
Change-Id: I196487da9c6a25441f3eaca3004d39b5f0220eb4