profile::wmcs::nfsclient: Move server mount path into yaml config
Previously there were a lot of assumptions here about where
'scratch' or 'project' or whatever were actually mounted on the NFS
server. Now that's actually specified in nfs-mounts.yaml.erb.

This is made easy by the fact that mount_nfs_volume.rb
already returned a value from the yaml file in the
'true' case.

Now that nfs-exportd doesn't know about 'home' vs 'project'
mounts, it will produce more granular export files, one
for each mount rather than one for the parent dir.

This refactor is because I'm in the process of moving a bunch
of these mounts onto NFS volumes and I want to specify where
the volumes are in one place rather in various places in
puppet and python code.

Change-Id: Ia2768cdc298ac60c29e9b467742d8f1611f0cbde