labstore nfs-exportd.py: Don't require public exports
Previously we regarded a lack of public or private exports
as an exceptional case and didn't update exports to avoid
breaking something. This made sense because this server
used to always have both.

Now that 'scratch' is on a VM server, labstores don't have
public exports anymore, so that meant we never updated
anything. Instead, narrow the exceptional case to a lack of
any exports at all.

Bug: T301280
Change-Id: I5716dd43582cf4911edc23c5c9baef5ea413d3b8