durum: add support for IPv6
As a continuation of adding IPv6 support to Wikidough, enable IPv6
support in bird/anycast-hc for durum by setting the relevant hiera
configuration and building on the work in f1f02470.

This commit also updates the type Service_ips to handle IPv6 addresses
and further updates the nginx configuration template to listen on those
addresses.

Bug: T301165
Hosts: O:durum
Change-Id: I1422809d7d55f40336bff1027ff95fcae323f6a0