rancid: rm .placeholder, ensure config dir exist, avoid puppet flap
As described in this old ticket from 2018 rancid is deleting the
file /var/lib/rancid/core/configs/.placeholder every hour and
then puppet recreates it afterwards.

This leads to puppet flapping which can STILL be seen on
https://puppetboard.wikimedia.org/node/netmon1002.wikimedia.org
since then.

Suggested fix here is to remove the .placeholder file and
ensure the "configs" subdirectory is a directory directly
instead of expecting the "recurse => remote" to do it
because of the .placeholder file.

Bug: T211459
Change-Id: Iecc199e1ee3a263ef47f202607ac72420af94bfa