gitlab: avoid $realm check, simplify ferm rules
As a follow-up to I0eed82cb28ef9c try to avoid the $realm checks
and simplify code by moving the value for the ferm drange over to Hiera
and let hat handle the difference between prod and cloud.

Realm checks are a bit discouraged from a coding style perspective.

Change-Id: I517f1a51b932b933e4ae42ee5a92db32d433b2fc