toolforge: redirect legacy ru_monuments to ru-monuments
The `ru_monuments` tool was made at a time when we did not require that
tool account names be RFC 1035 compliant "labels". This restriction
comes from our mapping of tool account names to Kubernetes namespaces.
The business logic for the ru_monuments tool has been migrated to the RFC
1035 compliant ru-monuments tool. This redirect will help them migrate
their users to the new tool. The redirect cannot be done from the former
tool itself due to the Kubernetes namespace issue associated with its
name.

Bug: T301720
Change-Id: I92072cab83b7bd27118bdfaac82b882000d72315