Add CR router IPv6 loopbacks to Bird config for esams
Current BIRD configuration for Anycast will configure itself to peer
with IP addresses passed in a list of neighbors.  At current live POPs,
i.e. excepting drmrs, Anycast nodes peer directly with the loopback
IPs of the CRs (as opposed to an IP on a particular private subnet).

The Bird templates are already designed to parse the list and build an
appropriate v4 and v6 configuration based on filtering this list and
determining the address family of each entry.

For drmrs current Bird templates also work fine as without this list
it will peer with default GW on connected Vlan, which works with
routed access layer.

Bug: T301165
Change-Id: I1e255b7d577d315b5cb4ff94f6b471e3c6390ffb