Remove ordered_yaml function
Prior to Ruby 1.9, a hash's order was not guaranteed resulting in
variable yaml output when converting hashes to yaml. As a consequence
ordered_yaml was added to ensure serialized yaml was ordered
consistently. Ruby 1.9 ensures hashes are ordered by their insertion
order. With this change in place we can drop ordered_yaml and use
stdlib's to_yaml. The output and order of to_yaml does differ from
ordered_yaml, so there will be a diff on the first puppet run, but
subsequent puppet runs should be diffless.

Hosts: P:base::production
Change-Id: I170466b83c8e993c00c583159a355fd59742175d