conftool: add request-actions / request-patterns
We want to be able to quickly enable/disable "abuse filters" at the edge
caching layer. To do this, we want to source such rules from etcd.

These rules, that we plan to store in "request-actions", will be a
combination (using AND or OR logic) of request patterns and of ipblocks,
that should be statically defined by us.

So we could ban all requests to search anything on cebwiki, or a specific
cache-busting query parameter for requests to enwiki only, and so on.

This patch is just the first in a series; a next one will need to introduce
the functionality in varnish vcl.

In the future, we can expand our ability to observe patterns by feeding
the same rules to the analytics real time analysis pipeline.

Change-Id: Ie4b1fc83863c29436291b0b60164e2875e396b77