prometheus: add heartbeat collection on mysqld_exporter
We're interested in scraping this information for ToolsDB, and there's
not an easy way to make the list of enabled exporters configurable. The
relevant grants should be deployed everywhere (at least according to the
grants .sql files on this repository), and I can't think of any harm
scraping this elsewhere too.

Change-Id: I0806a4147ede54e73f02c87b21393aec1e46f9a5