Add new Eqiad row E-F elastic servers to puppet with basic role
Eqiad rows (E and F) have been set up with power and networking,
which gets us to the stage we can now test imaging hosts and vlidate
that our normal deploy process works ok from the new cage.

The network setup in these new rows is a significant departure from
the configuration on other rows, so we need to test everything
thouroughly (see T299758).

Creating this change to enable some of the racked elastic servers
with the 'insetup' role, to validate DHCP, imaging process etc.
works as expected.  The search team can later modify this to move
them to the correct elasticsearch role as required.

Bug: T299609
Change-Id: I55d63c285310fe7cb5510d4ea5432262a02dbd51