Add new Eqiad private and analytics subnets to dhcp.conf
Trying to image a server in the new Eqiad cage I realised it wasn't
working as there was no entry for the new per-rack subnets on our DHCP
servers.

This adds the required entries for the net subnets which should cause
our install servers to respond to requests from hosts on these
networks.

Bug: T299758
Change-Id: I87a9800ac0bfdfb3afb91430e974011a6ebe1e92