run_ci_locally.sh: add podman support
Add the ability to run the CI containers with podman, rather than
docker. Also, when debugging with INTERACTIVE=yes the environment
variable ZUUL_REF was not set causing the run.sh script to fail. Merge
the duplicate args between the non-interactive & interactive modes in an
attempt to avoid future drift.

Change-Id: I409397473ab6dcdd5061935d2877080af59144ba