icinga: fix "legal check" monitoring footer HTML on en.wikipedia
This is Icinga monitoring for changes in the license information in the
HTML footer on en.wikipedia, called "check_legal_html" that
has existed for a long time and is still actively used.

An alert was created by this change:

https://en.wikipedia.org/w/index.php?title=MediaWiki:Wikimedia-copyright&diff=1072378717&oldid=861624479&diffmode=source

which was only changed a link label and is approved by legal.

A bit of a false positive but monitoring works as intended
and should continue to do so in the future, see ticket for that.

A bit of an issue here is that due to T253733 there is no email
sent to legal directly anymore as it used to happen in the past.

So that is why SRE had to see this on IRC or web and create tickets
to ping legal via Phabricator.

Bug: T302045
Change-Id: Id4e0644d3aa49a03bc86aa0979ecde2d5aab1128