P:mail::mx: add max_runtime_seconds to systemd_timer
Add a max runtime to the systemd timer of 30 minutes.  the timer should
run in much less then this however it only runs every hour so this
should be perfectly safe and conservative.

This should cause the service to be in an error state which in turn will
cause the icinga systemd status alert to trigger

Hosts: P:mail::mx
Change-Id: Ib8cfd0c88f4a9a25baaaef68d197bf72d223ad2b