Request high-entropy Sec-CH-UA* client hints
The User-Agent request header is useful for both community stewards when
attempting to identify misuse in the projects [0] and for Wikimedia
Foundation's public analytics and internal instrumentation.

We first need to collect low- and high-entropy client hints in order to
determine whether a combination of the two is an effective replacement
for the soon-to-be deprecated User-Agent request header.

We achieve the above by doing the following:

If the request was made over a secure transport, then set the Accept-CH
response header such that the following high-entropy client hints [0][1]
may be sent with subsequent requests by the UA:

- Sec-CH-UA-Arch
- Sec-CH-UA-Bitness
- Sec-CH-UA-Full-Version-List
- Sec-CH-UA-Model
- Sec-CH-UA-Platform-Version

Further, set the Permissions-Policy response header such that those
hints may also sent with requests to intake-analytics.wikimedia.org.

Finally, add the values of those high-entropy client hints to the
webrequest varnishkafka logs. Note well that the Sec-CH-UA,
Sec-CH-UA-Mobile, and Sec-CH-UA-Platform low-entropy client hints were
added in I2d18e3b9cfb.

[0] https://phabricator.wikimedia.org/T257893
[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Client_hints#user-agent_client_hints
[2] The Sec-CH-UA* client hints token names _not_ listed in
    https://wicg.github.io/client-hints-infrastructure/#low-entropy-table

Bug: T301238
Change-Id: Iffab38033fce29e6ca6743dabef496bfcdbb2913