wmcs-cinder-backup-manager: end with error if any backup failed
Currently we are ignoring any errors, and this ended up leaking
snapshots and delaying considerably the debugging of the current
issues due to the systemd service being considered up and running.
This makes sure that the systemd service will be in error state
whenever there's errors trying to backup anything, this might create
some alert that will go away, but given that we are not paging on them
better be safe than sorry.

Possible improvements include:
* Tweak systemd unit to restart a few times before giving up.
* Use an external file/db to keep track of the retries, and fail only
  if it fail the N previous times.
* Modify the script to run continuously, and keep the retry logic
  inside it.

Bug: T302299
Change-Id: Ifa963fedbb7de5f6edd33b0015111da813247dbe
Signed-off-by: David Caro <dcaro@wikimedia.org>