prometheus: ditch automatic icmp probes for service::catalog
As emerged in related task and T302265 the icmp probes are hitting LVS
rate limits. The probes are for historical reasons and to mimic Icinga
behaviour/limitations (services need to be attached to a (pingable)
host). Ditch the icmp probes altogether since L4+ probe failures have a
higher signal to noise ratio anyways.

Bug: T291946
Change-Id: I3b6e774548c6d2c73aa3ed9cab14a73a419f5a30