query_service: Repair oauth activation
While cleaning up the way blazegraph is started in Iae729082fb
the variable indicating if oauth is to be used was removed from
query_service::blazegraph, but was still needed by the
blazegraph-default.erb template. Bring back the value as a
boolean flag.

Bug: T302526
Hosts: wcqs1001.eqiad.wmnet,wcqs2002.codfw.wmnet
Change-Id: I5a3aa9f37ec3e9ed2f7dd2cd324b232fe5465ef8