query_service: pass cookies on to blazegraph
When using oauth these cookies are used as part of the request
logging process. While this could be conditional on oauth it
didn't seem like the extra variation in configurations would
be useful. There should be no harm passing the cookies on to wdqs
as well to be ignored.

Bug: T293462
Hosts: wcqs1001.eqiad.wmnet, wdqs1012.eqiad.wmnet
Change-Id: Ieb3cef462a6ce07e53fcd96ef8e41bee80c2c20a