wikistats: move repo from operations/debs on Gerrit to cloud on Gitlab
This repo once was an actual Debian package and at that time
all debs were under operatiobs/debs in Gerrit.

But nowadays this is NOT a deb anymore and also this is not operations
but only used in a cloud VPS project.

Moving it from Gerrit to Gitlab, imported with history and under
a new location that does not involve those misleading names anymore.

Fixes a long-time TODO and Ibf645988766511efe will make it possible
to use the puppet git::clone class to clone from Gitlab.

Change-Id: I857ebf36384a76da15bed67d986c8ba538e36c15