Restrict filesystem_avail_bigger_than_size check to Stretch
Most of the reports on the internet about negative free space on XFS
seem to occur around the same time frame. Then a commit was made to 4.18
which fixes, xfs: don't trip over negative free space in
xfs_reserve_blocks. I wasn't able to find any reports after this commit,
so I think there is a decent chance this commit fixed it or another one
did around the same time. Since we still have the workaround in place
and since all buster and greater servers are running kernels newer than
4.18 pin this only to stretch hosts.

Bug: T302687
Change-Id: Iebbe28cf949965d28a9787775e529d04acf77a01