query_service: Include scheme and host in X-redirect-url
This is providing `X-redirect-url: /foo/bar` to the backend. The
backend ends up interpreting this as http.  There might be something
to correct on that side so it knows it's in an https context, but
we can correct it from this side by always providing the https scheme.

From the user perspective this should be removing an extra round trip
which has to redirect the user back from http to https after auth.

There was also potential that the session cookie, marked Secure
HttpOnly and only provided over https, would cause a redirect loop,
but because the http->https bounce happens outside the wcqs instances
that only happens while debugging.

Hosts: wdqs1011.eqiad.wmnet, wcqs1001.eqiad.wmnet
Change-Id: Id9d030ea74f1f01e06349b9d84aad688d18013e5